git that hook nonsense outta here!
